Download state data if newest data bigger in size than latest download

+++ download-raw-data.r
@@ -63,7 +63,8 @@ if(length(previousData)>0){
   )
   
   print("Comparing latest state data to most recently downloaded data file.")
-  if(file.size(previousData)==file.size(tmp)){
+  notDownloadingHospitalBedData <- !grepl("HOSPITALS_esri",args[2],fixed=T)
+  if(file.size(previousData)>=file.size(tmp) && notDownloadingHospitalBedData){
     print("State's latest data has not been changed.")
     file.remove(tmp)
     stop(1)
